Split liquidity supportability
